Don't #define yyparse when using bison/flex
With bison/flex, the more modern way to allow multiple yacc/lex
grammars to be linked into a single executable is to use the

        bison: -p, --name-prefix=,PREFIX/, -Dapi.prefix=PREFIX
        flex: -Pprefix, --prefix=STRING

options, these take care of renaming all the machine-generated
global entry points, some of which are new.  When these options are
used "yylex", "yyparse", ... are already defined and our (potentially
incomplete) attempt to do the same conflicts with the "right" new
way to handle this.  This patch gets us out of the way when the job
has already been taken care of by the parser-generator.

Originally by Christos Zoulas.