Fix various iprop issues, extending the protocol
This commit extends the protocol in various ways, with what should be
graceful rolling upgrade semantics (you should be able to upgrade the
master and the slaves in any order and not all at once).

The master and slave now identify state not by a single uint32_t version
number, but also by the timestamp at which an iprop log entry was
recorded at that version.  This makes reinitialization of iprop safe.

This means:

 - protocol messages were extended:

    - I_HAVE
    - TELL_YOU_EVERYTHING (which appears in the dump)
    - NOW_YOU_HAVE   (which also appears in the dump)

 - master datastructures were extended

 - the dump format was extended in that the TELL_YOU_EVERYTHING /
   NOW_YOU_HAVE brackets were extended

 - a previous commit ensures that the log always has at least two
   entries: the uber record and a first record, which means that we
   always have at least one record whose version and timestamp can be
   used to help establish identify synchronization state