lib/krb5: restore krb5_context_data field ordering to match 7.x
The struct krb5_context_data field can be passed to plugins
which might not be aware of the new structure layout.  To
reduce the risk of data corruption, fields must never be
removed, modified or reordered.  Old unused fields are marked
deprecated and new fields are appended to the end of the structure.

This change moves two fields that were added to master which
are not present in Heimdal 7.x:

fe43be85587f834266623adb0ecf2793d212a7ca added config_include_depth.

5b39bd7c1d5447e544498962e93ac06c86f9d1f2 added no_ticket_store.

Change-Id: I28b157e128732324972c99b246a93a828bc077c1