s4u2self: fix service code in cross realm trust
When impersonating a user from another realm, we first need
to obtain a TGT to TGS in that user's realm.
Only then we can ask for a ticket for self and follow back
referrals to our own realm.
See MS-SFU sections 3.1.5.1.1.1 and 4.2

In case we fail to obtain that krbtgt, fall back to asking
from the service's own realm, as it may be the same realm
using different name such as netbios or lower realm.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>